+++ src/api/fetchImgixMetadata.ts
@@ -6,7 +6,7 @@ import ImgixClient from 'imgix-core-js';
 import * as t from 'io-ts';
 import { createLogger, trace } from '../common/log';
 import { withCache } from '../common/cache';
-import { fetchJSON } from '../modules/gatsby-source-url/utils';
+import { fetchJSON } from '../common/utils';
 
 const log = createLogger('fetchImgixMetadata');
